backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
torch
torchvision
ultralytics
faiss-cpu
numpy
pydantic
